Remove: remove graphql label from folders
