+++ bigbluebutton-html5/imports/ui/components/nav-bar/recording-indicator/component.tsx
