+++ bigbluebutton-html5/imports/ui/components/nav-bar/talking-indicator/queries.ts
