+++ bigbluebutton-html5/imports/ui/components/nav-bar/talking-indicator/service.ts
