+++ bigbluebutton-html5/imports/ui/components/presence-adapter/component.tsx
