+++ bigbluebutton-html5/imports/ui/components/waiting-users/component.tsx
