+++ bigbluebutton-html5/imports/ui/components/waiting-users/guest-items/guestPendingUser.tsx
