+++ bigbluebutton-html5/imports/ui/components/waiting-users/guest-items/noPendingGuestUser.tsx
