+++ bigbluebutton-html5/imports/ui/components/waiting-users/styles.ts
